Split crypto functions out of Util class so they don't have to be included everywhere
